Generate the locations from Python.
Locations are specified in Python, so that Python scripts may access the xtx
shortcuts.  Using the Python representation, locations.py can generate the
existing xtx/locations.xtx file to be almost identical (save for whitespace).
The places in which locations.xtx changes with this commit are for the better,
as it makes it more consistent.  The removal of whitespace makes it harder for a
human to parse, but it's still easy to parse.

TODO:
 - Perform justification so that locations.xtx is easy to parse.