Allow conferences to appear under multiple book titles.
An example problematic title:

    Proceedings of the 1976 ACM SIGMOD International Conference on Management of
    Data, Washington, D.C., June 2-4, 1976

We desire to just call it:

    SIGMOD International Conference on Management of Data

We can't just use the fact that it is called 'conf/sigmod/*' because workshops
often use the same key prefix as their host conference.  I don't want workshop
papers appearing to come from conferences.

Our solution is to loosely look for "long name" in "booktitle".